chore: rename observe to collect
