+++ panda_agi/train/collect.py
@@ -40,7 +40,7 @@ def get_available_proxies():
         
     return proxies
 
-class observe(ContextDecorator):
+class collect(ContextDecorator):
     def __init__(self, model_name: Optional[str] = None, providers: Optional[List[str]] = None, debug: Optional[bool] = False):
         self.model_name = model_name
         available = get_available_proxies()
